refactor(RemoteAccessor): abstraction into abstract class for RemoteAccessor and RemoteAccessorMenu
